+++ examples/nuxt/colada.options.ts
@@ -1,0 +1,5 @@
+import type { PiniaColadaOptions } from '@pinia/colada'
+
+export default {
+  // Options here
+} satisfies PiniaColadaOptions