+++ src/codegen/generateDTS.ts
@@ -1,3 +1,5 @@
+import { ts } from '../utils'
+
 export function generateDTS({
   routesModule,
   routeNamedMap,
@@ -6,7 +8,7 @@ export function generateDTS({
   routesModule: string
   routeNamedMap: string
 }) {
-  return `
+  return ts`
 /* eslint-disable */
 /* prettier-ignore */
 // @ts-nocheck