+++ src/codegen/generateRouteRecords.ts
@@ -1,3 +1,4 @@
+import { getLang } from '@vue-macros/common'
 import type { TreeNode } from '../core/tree'
 import { ImportsMap } from '../core/utils'
 import { type ResolvedOptions } from '../options'
@@ -24,6 +25,27 @@ ${node
   .map((child) => generateRouteRecord(child, options, importsMap, indent + 1))
   .join(',\n')}
 ]`
+  }
+
+  const definePageDataList: string[] = []
+
+  if (node.hasDefinePage) {
+    for (const [name, filePath] of node.value.components) {
+      const pageDataImport = `_definePage_${name}_${importsMap.size}`
+      definePageDataList.push(pageDataImport)
+      const lang = getLang(filePath)
+      importsMap.addDefault(
+        // TODO: apply the language used in the sfc
+        `${filePath}?definePage&` +
+          (lang === 'vue' ? 'vue&lang.tsx' : `lang.${lang}`),
+        pageDataImport
+      )
+    }
+
+    // extra indent to add `_mergeRouteRecord()`
+    if (definePageDataList.length > 0) {
+      indent++
+    }
   }
 
   const startIndent = ' '.repeat(indent * 2)
@@ -75,25 +97,14 @@ ${indentStr}],`
   }${formatMeta(node, indentStr)}
 ${startIndent}}`
 
-  if (node.hasDefinePage) {
-    const definePageDataList: string[] = []
-    for (const [name, filePath] of node.value.components) {
-      const pageDataImport = `_definePage_${name}_${importsMap.size}`
-      definePageDataList.push(pageDataImport)
-      importsMap.addDefault(
-        // TODO: apply the language used in the sfc
-        `${filePath}?definePage&vue&lang.tsx`,
-        pageDataImport
-      )
-    }
-
-    if (definePageDataList.length) {
-      importsMap.add('unplugin-vue-router/runtime', '_mergeRouteRecord')
-      return `  _mergeRouteRecord(
+  if (definePageDataList.length > 0) {
+    // remove one tab
+    const mergeCallIndent = startIndent.slice(2)
+    importsMap.add('unplugin-vue-router/runtime', '_mergeRouteRecord')
+    return `${mergeCallIndent}_mergeRouteRecord(
 ${routeRecord},
-  ${definePageDataList.join(',\n')}
-  )`
-    }
+${definePageDataList.map((s) => startIndent + s).join(',\n')}
+${mergeCallIndent})`
   }
 
   return routeRecord