+++ src/data-loaders/entries/index.ts
@@ -4,12 +4,20 @@ export type {
   UseDataLoaderResult,
   DataLoaderContextBase,
   DataLoaderEntryBase,
-  DefineDataLoaderOptionsBase,
+  DefineDataLoaderOptionsBase_LaxData,
+  DefineDataLoaderOptionsBase_DefinedData,
   DefineLoaderFn,
+  // deprecated
+  DefineDataLoaderOptionsBase,
 } from '../createDataLoader'
+export { toLazyValue } from '../createDataLoader'
 
 // new data fetching
-export { DataLoaderPlugin, NavigationResult } from '../navigation-guard'
+export {
+  DataLoaderPlugin,
+  NavigationResult,
+  useIsDataLoading,
+} from '../navigation-guard'
 export type {
   DataLoaderPluginOptions,
   SetupLoaderGuardOptions,
@@ -27,5 +35,7 @@ export {
   currentContext,
 } from '../utils'
 
-// expose all symbols that could be use by loaders
+// expose all symbols that could be used by loaders
 export * from '../meta-extensions'
+
+export type { TypesConfig, ErrorDefault } from '../types-config'