+++ src/data-loaders/navigation-guard.spec.ts
@@ -2,7 +2,11 @@
  * @vitest-environment happy-dom
  */
 import { App, createApp, defineComponent } from 'vue'
-import { DefineDataLoaderOptions, defineBasicLoader } from './defineLoader'
+import {
+  DefineDataLoaderOptions_LaxData,
+  DefineDataLoaderOptions_DefinedData,
+  defineBasicLoader,
+} from './defineLoader'
 import {
   afterAll,
   afterEach,
@@ -20,6 +24,7 @@ import {
   DataLoaderPlugin,
   NavigationResult,
   DataLoaderPluginOptions,
+  useIsDataLoading,
 } from 'unplugin-vue-router/data-loaders'
 import { mockPromise } from '../../tests/utils'
 import {
@@ -30,7 +35,9 @@ import type { NavigationFailure } from 'vue-router'
 
 function mockedLoader<T = string | NavigationResult>(
   // boolean is easier to handle for router mock
-  options?: DefineDataLoaderOptions
+  options:
+    | DefineDataLoaderOptions_LaxData
+    | DefineDataLoaderOptions_DefinedData = {}
 ) {
   const [spy, resolve, reject] = mockPromise<T, unknown>(
     // not correct as T could be something else
@@ -333,6 +340,35 @@ describe('navigation-guard', () => {
     'does not call commit for a loader if the navigation is canceled by another loader'
   )
 
+  it('sets isDataLoading within a navigation', async () => {
+    const { app } = setupApp({ isSSR: false })
+    const isGloballyLoading = app.runWithContext(() => useIsDataLoading())
+    expect(isGloballyLoading.value).toBe(false)
+
+    const router = getRouter()
+    const l1 = mockedLoader()
+    const l2 = mockedLoader()
+    router.addRoute({
+      name: '_test',
+      path: '/fetch',
+      component,
+      meta: {
+        loaders: [l1.loader, l2.loader],
+      },
+    })
+
+    router.push('/fetch')
+    await vi.runOnlyPendingTimersAsync()
+    expect(isGloballyLoading.value).toBe(true)
+
+    l1.resolve()
+    await vi.runAllTimersAsync()
+    expect(isGloballyLoading.value).toBe(true)
+    l2.resolve()
+    await vi.runAllTimersAsync()
+    expect(isGloballyLoading.value).toBe(false)
+  })
+
   describe('signal', () => {
     it('aborts the signal if the navigation throws', async () => {
       setupApp({ isSSR: false })
@@ -580,7 +616,7 @@ describe('navigation-guard', () => {
       expect(router.currentRoute.value.fullPath).toBe('/fetch')
     })
 
-    it('handle global expected errors even when rejected by local errors', async () => {
+    it('local errors completely override global ones', async () => {
       setupApp({
         isSSR: false,
         // global only accepts CustomError
@@ -605,7 +641,8 @@ describe('navigation-guard', () => {
       await vi.runOnlyPendingTimersAsync()
       l1.reject(new CustomError())
       await router.getPendingNavigation().catch(() => {})
-      expect(router.currentRoute.value.fullPath).toBe('/fetch')
+      // the navigation was not aborted
+      expect(router.currentRoute.value.fullPath).toBe('/')
     })
   })
 })